+++ stripe_terminal/android/src/main/kotlin/mek/stripeterminal/Utils.kt
@@ -8,3 +8,7 @@ private val mainThread = Handler(Looper.getMainLooper())
 fun runOnMainThread(body: () -> Unit) {
     mainThread.post(body)
 }
+
+fun microsecondsToSeconds(value: Long): Int {
+    return (value * 1000000).toInt()
+}